Move canonicalize to base
